[refactor] MemberService, AuthService 분리
